v2_admin/requirements.txt
python-telegram-bot==20.7
flask==3.0.3
python-dotenv==1.0.1
waitress==3.0.0
Werkzeug==3.0.3
